ix some minor issues
