+++ .github/workflows/api-test.yml
@@ -92,8 +92,6 @@ jobs:
             class: org.apache.dolphinscheduler.api.test.cases.LoginAPITest
           - name: Tenant
             class: org.apache.dolphinscheduler.api.test.cases.TenantAPITest
-          - name: TenantNew
-            class: org.apache.dolphinscheduler.api.test.cases.TenantAPITestNew
     env:
       RECORDING_PATH: /tmp/recording-${{ matrix.case.name }}
     steps:
