+++ dolphinscheduler-api-test/dolphinscheduler-api-test-case/src/test/java/org/apache/dolphinscheduler/api.test/pages/security/tenant/TenantPageAPI.java
@@ -19,15 +19,14 @@ package org.apache.dolphinscheduler.api.test.pages.security.tenant;
 
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
-import org.apache.dolphinscheduler.api.test.core.common.FormParam;
-import org.apache.dolphinscheduler.api.test.entity.PageParamEntity;
+import org.apache.dolphinscheduler.api.test.core.common.Constants;
+import org.apache.dolphinscheduler.api.test.entity.PageRequestEntity;
 import org.apache.dolphinscheduler.api.test.core.common.RequestMethod;
 import org.apache.dolphinscheduler.api.test.pages.Route;
 import org.apache.dolphinscheduler.api.test.pages.security.tenant.entity.TenantRequestEntity;
 import org.apache.dolphinscheduler.api.test.utils.RestResponse;
 import org.apache.dolphinscheduler.api.test.utils.Result;
 
-import static io.restassured.RestAssured.given;
 
 public class TenantPageAPI implements ITenantPageAPI {
     private final RequestSpecification reqSpec;
@@ -41,20 +40,20 @@ public class TenantPageAPI implements ITenantPageAPI {
 
     @Override
     public RestResponse<Result> createTenant(TenantRequestEntity tenantRequestEntity) {
-        return toResponse(RestRequestByRequestMap(getRequestNewInstance().spec(reqSpec), sessionId,
+        return toResponse(restRequestByRequestMap(getRequestNewInstance().spec(reqSpec), sessionId,
                 tenantRequestEntity.toMap(), Route.tenants(), RequestMethod.POST));
     }
 
     @Override
     public RestResponse<Result> updateTenant(TenantRequestEntity tenantUpdateEntity, int tenantId) {
-        return toResponse(RestRequestByRequestMap(getRequestNewInstance().spec(reqSpec), sessionId,
+        return toResponse(restRequestByRequestMap(getRequestNewInstance().spec(reqSpec), sessionId,
                 tenantUpdateEntity.toMap(), Route.tenants(tenantId), RequestMethod.PUT));
     }
 
     @Override
-    public RestResponse<Result> getTenants(PageParamEntity pageParamEntity) {
+    public RestResponse<Result> getTenants(PageRequestEntity pageParamEntity) {
         Response resp = getRequestNewInstance().spec(reqSpec).
-                cookies(FormParam.SESSION_ID.getParam(), sessionId).
+                cookies(Constants.SESSION_ID_KEY, sessionId).
                 params(pageParamEntity.toMap()).
                 when().get(Route.tenants());
         return toResponse(resp);
@@ -63,7 +62,7 @@ public class TenantPageAPI implements ITenantPageAPI {
     @Override
     public RestResponse<Result> getTenantsListAll() {
         Response resp = getRequestNewInstance().spec(reqSpec).
-                cookies(FormParam.SESSION_ID.getParam(), sessionId).
+                cookies(Constants.SESSION_ID_KEY, sessionId).
                 when().get(Route.tenantsList());
         return toResponse(resp);
     }
@@ -71,7 +70,7 @@ public class TenantPageAPI implements ITenantPageAPI {
     @Override
     public RestResponse<Result> verifyTenantCode(String tenantCode) {
         Response resp = getRequestNewInstance().spec(reqSpec).
-                cookies(FormParam.SESSION_ID.getParam(), sessionId).
+                cookies(Constants.SESSION_ID_KEY, sessionId).
                 queryParam("tenantCode", tenantCode).
                 when().get(Route.tenantsVerifyCode());
         return toResponse(resp);
@@ -80,7 +79,7 @@ public class TenantPageAPI implements ITenantPageAPI {
     @Override
     public RestResponse<Result> deleteTenantById(int tenantId) {
         Response resp = getRequestNewInstance().spec(reqSpec).
-                cookies(FormParam.SESSION_ID.getParam(), sessionId).
+                cookies(Constants.SESSION_ID_KEY, sessionId).
                 when().delete(Route.tenants(tenantId));
         return toResponse(resp);
     }
