+++ dolphinscheduler-api-test/dolphinscheduler-api-test-case/src/test/java/org/apache/dolphinscheduler/api.test/pages/security/tenant/entity/TenantRequestEntity.java
@@ -18,9 +18,46 @@
 
 package org.apache.dolphinscheduler.api.test.pages.security.tenant.entity;
 
-public class TenantRequestEntity extends TenantBaseEntity {
+import org.apache.dolphinscheduler.api.test.base.AbstractBaseEntity;
+
+public class TenantRequestEntity extends AbstractBaseEntity {
     private int id;
+
+    private String tenantCode;
+
+    private String description;
+
+    private int queueId;
+
+    public int getId() {
+        return id;
+    }
+
     public void setId(int id) {
         this.id = id;
+    }
+
+    public String getTenantCode() {
+        return tenantCode;
+    }
+
+    public void setTenantCode(String tenantCode) {
+        this.tenantCode = tenantCode;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public int getQueueId() {
+        return queueId;
+    }
+
+    public void setQueueId(int queueId) {
+        this.queueId = queueId;
     }
 }
